requirements.txt
cachetools==5.3.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.9" and python_version < "4"
google-api-core==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-bigquery==3.6.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version < "4.0" and python_version >= "3.9"
grpcio-status==1.51.3 ; python_version < "4.0" and python_version >= "3.9"
grpcio==1.51.3 ; python_version < "4.0" and python_version >= "3.9"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
libtorrent==2.0.7 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.22.0 ; python_version < "4.0" and python_version >= "3.9"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
torrentp==0.1.5 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4"
